Query all builds and pick the latest ourselves
The "latest" endpoint from wago is the latest by _date_, which would
include chinese servers which are way behind the rest.

This means that this tool officially does not support addons for the
chinese market, but that could change if there's demand for it.

For now this is just the easier path, but it assumes the list is sorted
correctly by version.